requirements.txt
fastmcp>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0
typing-extensions>=4.8.0
pandas>=2.0.0
a2a-sdk>=0.1.0
uvicorn>=0.24.0
litellm>=1.30.0
google-adk>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5 